requirements.txt
# =============================================================================
# GRUPONOS MELTANO NATIVE - PYTHON REQUIREMENTS
# =============================================================================

# Core Meltano framework
meltano>=3.0.0,<4.0.0

# Oracle database support
oracledb>=2.0.0
cx-Oracle>=8.3.0

# dbt with Oracle adapter
dbt-core>=1.8.0,<2.0.0
dbt-oracle>=1.8.0,<2.0.0

# Data processing and utilities
pandas>=2.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# HTTP client for APIs
httpx>=0.28.0
requests>=2.31.0

# Configuration and environment
python-dotenv>=1.0.0
click>=8.1.0

# Monitoring and logging
loguru>=0.7.0
structlog>=23.0.0

# Data validation and quality
great-expectations>=0.18.0

# Performance and optimization
orjson>=3.9.0
pyarrow>=14.0.0

# Optional: Apache Airflow for orchestration
# Uncomment if using --with-airflow setup
# apache-airflow>=2.9.0
# apache-airflow-providers-oracle>=3.10.0

# Development and testing dependencies
pytest>=8.0.0
pytest-cov>=5.0.0
ruff>=0.8.0
mypy>=1.13.0
pre-commit>=4.0.0

# Additional utilities
rich>=13.9.0
typer>=0.12.0
croniter>=2.0.0